Fix some UB and warnings

+++ include/game/armem.h
@@ -6,16 +6,16 @@
 #include "game/data.h"
 
 void HuARInit(void);
-u32 HuARMalloc(u32 size);
-void HuARFree(u32 amemptr);
+uintptr_t HuARMalloc(u32 size);
+void HuARFree(uintptr_t amemptr);
 void HuAMemDump(void);
-u32 HuAR_DVDtoARAM(u32 dir);
-u32 HuAR_MRAMtoARAM(s32 dir);
-u32 HuAR_MRAMtoARAM2(void *dir_ptr);
-void HuAR_ARAMtoMRAM(u32 amemptr);
-void *HuAR_ARAMtoMRAMNum(u32 amemptr, s32 num);
+uintptr_t HuAR_DVDtoARAM(u32 dir);
+uintptr_t HuAR_MRAMtoARAM(s32 dir);
+uintptr_t HuAR_MRAMtoARAM2(void *dir_ptr);
+void HuAR_ARAMtoMRAM(uintptr_t amemptr);
+void *HuAR_ARAMtoMRAMNum(uintptr_t amemptr, s32 num);
 s32 HuARDMACheck(void);
-u32 HuARDirCheck(u32 dir);
+uintptr_t HuARDirCheck(u32 dir);
 void HuARDirFree(u32 dir);
 void *HuAR_ARAMtoMRAMFileRead(u32 dir, u32 num, HeapID heap);
